+++ kaithem/src/docs/changes.md
@@ -2,7 +2,9 @@ Change Log
 ----------
 ### 0.68.15
 - kaithem.web.serveFile now can serve a bytesIO object if mime and filename are provided.
-
+- Object Detection in NVRChannel!!!
+- Fix bug where deleting a tag point logger would not save.
+  
 ### 0.68.14
 - Fix bug where NVRChannel would carry over record sessions and thereby crash and not be able to recover from connection failure
 - Improve nvr time search